log original lifetimes as floats

+++ exomol2lida/process_dataset.py
@@ -392,7 +392,7 @@ class DatasetProcessor:
         if self.include_original_lifetimes and "tau" in self.states_header:
             if lumped_state not in self.states_map_lumped_to_tau:
                 self.states_map_lumped_to_tau[lumped_state] = []
-            self.states_map_lumped_to_tau[lumped_state].extend(df.tau)
+            self.states_map_lumped_to_tau[lumped_state].extend(df.tau.astype("float64"))
 
         # now calculate the lumped state attributes:
         # energy is calculated as the average of energies over the states with
